requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-openai
langchain-google-genai
langchain-postgres
pypdf
python-dotenv
psycopg[binary]
